requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
mutagen>=1.47.0
Pillow>=10.0.0
